requirements.txt
opencv_python==4.8.1.78
Flask==3.0.0
flask-sqlalchemy==3.1.1
watchdog==3.0.0